Reworked the code for variable cloning, and used the new code as the basis for adding variables (interface in vartable_ui.c, code in varchange.c).  One can now add a variable representing the brushed groups, or a variable of row numbers.  Cleaned up the code in ggobi-API.c to correspond.
Encountered a bug in sphering -- another variable needed to be
set to NULL when the window was closed.

Played a bit more with the resizing of the main panel (main_ui.c).

Added a bit of error-prevention to the code for removing
event handlers, to make sure we don't try to remove one that
isn't there any longer.