varstats table:   1. added a 'rename variable' dialog   2. grouped buttons differently along the bottom of the screen
hide/exclude panel:
  1. don't destroy the entire window when updating
  2. renamed a label

Fixed a bug in reading column labels from ascii file: it wasn't
handling blanks in variable names, because it was relying on
white space as the separator between fields (where the fields
can be var_name    minimum    maximum).  I'm now using '|' as
a field separator -- and using g_strstrip to remove leading
and trailing whitespace in the variable names, too, which is
nice.