Single-column data was being handled inconsistently -- ie, the second column was being added to some matrices and not to others. I'm not going to add columns at all for now.
I never noticed this bug before, so maybe it's a transient one,
but the values of the rulers were not being correctly initialized,
so I'm forcing them to be set when the scatterplot display is
initialized.

notes/debugging contains some advice for making gtk applications
stop when there's a gtk warning.