The basic data types in GGobi (ggobid, datad, displayd, splotd) are now classes in the Gtk scheme.
ggobid extends GtkObject as class GtkGGobi
datad  extends GtkObject as class GtkGGobiData
displayd extends GtkVBox as class GtkGGobiDisplay
   windowDisplayd extends GtkGGobiDisplay as GtkGGobiWindowDisplay
splotd  extends GtkDrawingArea as class GtkGGobiSPlot

One of the benefits of this is that we can now access these objects in
Rggobi using the Gtk bindings provided by the RGtk mechanism.  This
means that with no additional code, we get access to the
signals/callbacks defined by these objects.  We can also get at the
arguments/properties of these objects if there are any. (More to
come.)

The primary motivation is to allow extended plot types so that others
can provide these within plugins. This is the next step in this
addition.